feat: implement referral system with OTP verification

+++ src/main/java/bullbear/app/controller/NotificationController.java
@@ -28,7 +28,7 @@ public class NotificationController {
         }
 
         return notificationUtil.getUserNotifications(
-                currentUser.getId().intValue()
+                currentUser.getId()
         );
     }
 }
